Fix bug where client freezes on player vs player collision
This is due to the fact that the current coordinates are used for collision
detection rather than the future coordinates that the player requested. Thus,
by the time the server tells the player it can't move because there will be a
collision, the player has already crashed into another player and both of them
are dead(locked). A new method that can detect nearby players around specified
coordinates is added in order to solve this problem. Fix #2.