Change default game end condition
The end condition is changed so that the game finishes when one team collects
all of its team's points. Note that teams that see other team's food as their
own food will be at disadvantage. The rationale for this change is that once
all visible food is consumed, some teams are left with nothing to do.