Add feature to configure basic game settings by file
This adds a new class BarngaOnlineConfigsFromFile that inherits from
BarngaOnlineConfigsDefault and reads in a configuration file to make the basic
parameters configurable. Default values are used for values that are not
stated in the configuration file. Toml is used for the configuration file
since it is portable while maintaining readability. Fix #9.