Correctly handle world state at the end of game
With this commit, both the server and the client resets the world state. The
server resets the list of connected clients, teams, players, and points as
well as setting the gameStarted to false. The client clears its own
information such as ID and coordinates, the list of connected players, and the
list of food. This change allows back-to-back game plays. Fix #10.