Improve performance with many food in world
This commit changes the way food is managed. The world is divided into tiles,
and each food is stored into the corresponding 2D array that represents tiles.
When the player moves, only the food in the adjacent tiles is rendered,
decreasing the amount of food that needs to be iterated through. This change
drastically improves the number of food that the client can handle. The client
can handle a 30,000 x 30,000 world size with more than 20,000 food tokens with
almost no delay. However, with 40,000 food, the client gets very slow. Fix #21.