Allow back-to-back game playing
This change allows games to be played back-to-back. restartGame() method is
added to the WorldState class, which resets the teams, points, and food and
generates a new set. The clients are kept track now using SocketIOClient
objects as the keys instead of SocketAddress, since that makes it easier to
send data to the already-connected clients when restarting the game.