[BUGFIX] game finishes instantly on second round
This bug happens when not all teams are used (when team_assignment is larger
than the actual teams played) because food is generated only when players are
added to the team. If no player is added, then there will be no food for that
team, and thus the game end condition check passes right after the first move.